post-service/requirements.txt

fastapi
uvicorn
motor
pymongo
python-jose[cryptography]